requirements.txt
asgiref==3.7.2
Django==4.2.10
djangorestframework==3.14.0
psycopg2-binary==2.9.3
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.9.0
